Facades

+++ MantenimientoTPI-ejb/src/main/java/ues/fmoocc/ingenieria/tpi2018/Sessions/AbstractFacade.java
@@ -22,21 +22,59 @@ public abstract class AbstractFacade<T> {
 
     protected abstract EntityManager getEntityManager();
 
-    public void create(T entity) {
-        getEntityManager().persist(entity);
+    public T create(T entity) {
+        if (entity != null && getEntityManager()!=null) {
+            getEntityManager().persist(entity);
+            return entity;
+        }else{
+            return null;
+        }
     }
 
-    public void edit(T entity) {
-        getEntityManager().merge(entity);
+    public T edit(T entity) {
+        if (entity != null && getEntityManager()!=null) {
+            getEntityManager().merge(entity);
+            return entity;
+        }else{
+            return null;
+        }
     }
 
-    public void remove(T entity) {
-        getEntityManager().remove(getEntityManager().merge(entity));
+    public T remove(T entity) {
+        if (entity!=null && getEntityManager()!=null) {
+            getEntityManager().remove(getEntityManager().merge(entity));
+            return entity;
+        }else{
+            return null;
+        }
     }
 
     public T find(Object id) {
         return getEntityManager().find(entityClass, id);
     }
+    
+     public boolean crear(T entity){
+        
+        if (entity!= null && getEntityManager()!=null) {
+            return create(entity)==entity;
+        }
+        return false;
+    }
+    
+    public boolean modificar(T entity){
+        if (entity!= null && getEntityManager()!=null) {
+            return edit(entity)==entity;
+        }
+        return false;
+    }
+    
+    public boolean eliminar(T entity){
+        if (entity!= null && getEntityManager()!=null) {
+            return remove(entity)==entity;
+        }
+        return false;
+    }
+
 
     public List<T> findAll() {
         javax.persistence.criteria.CriteriaQuery cq = getEntityManager().getCriteriaBuilder().createQuery();
